+++ src/main/java/com/agrimeme/backend/controller/PostController.java
@@ -1,6 +1,8 @@
 package com.agrimeme.backend.controller;
 
 import com.agrimeme.backend.exceptions.ResourceNotFoundException;
+import com.agrimeme.backend.exceptions.BadRequestException;
+
 import com.agrimeme.backend.model.Post;
 import com.agrimeme.backend.model.VoteIdentity;
 import com.agrimeme.backend.model.Votes;
@@ -49,16 +51,18 @@ public class PostController {
     public Post createPost(@Valid @RequestBody Post post) {
     	Long userId = ((UserPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
     	return userRepository.findById(userId).map(user -> {
-            post.setUser(user);
+            post.setUserId(user.getId());
             return postRepository.save(post);
         }).orElseThrow(() -> new ResourceNotFoundException("UserId " + userId + " not found"));
     	
     }
-    
+    @RolesAllowed("ROLE_USER")
     @PutMapping("/posts/{postId}")
     public Post updatePost(@PathVariable Long postId, @Valid @RequestBody Post postRequest) {
-    	
+    	Long userId = ((UserPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
         return postRepository.findById(postId).map(post -> {
+            if(post.getUserId() != userId)
+                throw new BadRequestException("Unauthorized Request.");
             post.setTitle(postRequest.getTitle());
             post.setDescription(postRequest.getDescription());
             post.setImageUrl(postRequest.getImageUrl());
@@ -66,31 +70,32 @@ public class PostController {
         }).orElseThrow(() -> new ResourceNotFoundException("PostId " + postId + " not found"));
     }
     
+    @RolesAllowed("ROLE_USER")
     @PostMapping("/posts/{postId}/upvote")
-    public Post upvotePost(@PathVariable Long postId) {
-    	Long userId = ((UserPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
-    	VoteIdentity vi = new VoteIdentity(postId, userId);
-    	Optional<Votes> vote = votesRepository.findById(vi);
-    	if(vote.isPresent()) {
-    		return postRepository.findById(postId).map(post -> {
-    			Votes voteObj = vote.get();
-    			post.setVotes(post.getVotes() - voteObj.getValue());
-    			voteObj.setValue(1L);
-    			post.setVotes(post.getVotes() + voteObj.getValue());
-                return postRepository.save(post);
-            }).orElseThrow(() -> new ResourceNotFoundException("PostId " + postId + " not found"));
-    	}
-    	return postRepository.findById(postId).map(post -> {
+	public Post upvotePost(@PathVariable Long postId) {
+		Long userId = ((UserPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
+		VoteIdentity vi = new VoteIdentity(postId, userId);
+		Optional<Votes> vote = votesRepository.findById(vi);
+		if(vote.isPresent()) {
+			return postRepository.findById(postId).map(post -> {
+				Votes voteObj = vote.get();
+				post.setVotes(post.getVotes() - voteObj.getValue());
+				voteObj.setValue(1L);
+				post.setVotes(post.getVotes() + voteObj.getValue());
+	            return postRepository.save(post);
+	        }).orElseThrow(() -> new ResourceNotFoundException("PostId " + postId + " not found"));
+		}
+		return postRepository.findById(postId).map(post -> {
 			Votes voteObj = new Votes();
 			voteObj.setVoteIdentity(vi);
 			voteObj.setValue(1L);
 			votesRepository.save(voteObj);
 			post.setVotes(post.getVotes() + voteObj.getValue());
-            return postRepository.save(post);
-        }).orElseThrow(() -> new ResourceNotFoundException("PostId " + postId + " not found"));
-    	
-    }
-    
+	        return postRepository.save(post);
+	    }).orElseThrow(() -> new ResourceNotFoundException("PostId " + postId + " not found"));
+		
+	} 
+    @RolesAllowed("ROLE_USER")
     @PostMapping("/posts/{postId}/downvote")
     public Post downvotePost(@PathVariable Long postId) {
     	Long userId = ((UserPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
@@ -115,10 +120,13 @@ public class PostController {
         }).orElseThrow(() -> new ResourceNotFoundException("PostId " + postId + " not found"));
     	
     }
-    
+    @RolesAllowed("ROLE_USER")
     @DeleteMapping("/posts/{postId}")
     public ResponseEntity<?> deletePost(@PathVariable Long postId) {
+        Long userId = ((UserPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
         return postRepository.findById(postId).map(post -> {
+            if(post.getUserId() != userId)
+                throw new BadRequestException("Unauthorized Request.");
             postRepository.delete(post);
             return ResponseEntity.ok().build();
         }).orElseThrow(() -> new ResourceNotFoundException("PostId " + postId + " not found"));
